refactor(components): split SelectionPopup into modular composables and sub-components
